♻️ refactor : App.js 페이지별 컴포넌트 분리
