+++ src/page-layout/FolderPage/FolderPage.jsx
@@ -1,12 +1,15 @@
+import { AddLinkBar } from "ui/AddLinkBar/AddLinkBar";
 import "./FolderPage.css";
+import { SearchBar } from "ui/SearchBar";
+import { FolderContent } from "ui/FolderContent/FolderContent";
 
-export const FolderPage = ({ addLink, searchBar, content }) => {
+export const FolderPage = () => {
   return (
     <div className="FolderPage">
-      {addLink}
+      <AddLinkBar />
       <div className="FolderPage-items">
-        {searchBar}
-        {content}
+        <SearchBar />
+        <FolderContent />
       </div>
     </div>
   );
