+++ src/page-layout/SharedPage/SharedPage.jsx
@@ -1,12 +1,41 @@
+import { useGetFolder } from "data-access/useGetFolder";
 import "./SharedPage.css";
+import { FolderInfo } from "ui/FolderInfo";
+import { SearchBar } from "ui/SearchBar";
+import { CardList } from "ui/CardList";
+import { ReadOnlyCard } from "ui/ReadOnlyCard";
+import { getFolders } from "data-access/getFolders";
+import { useEffect, useState } from "react";
 
-export const SharedPage = ({ folderInfo, searchBar, cardList }) => {
+export const SharedPage = () => {
+  const [folders, setFolders] = useState();
+
+  const { data } = useGetFolder();
+  const { profileImage, ownerName, folderName } = data || {};
+
+  const handleLoadFolders = async () => {
+    const { data } = await getFolders();
+    setFolders(data);
+  };
+
+  useEffect(() => {
+    handleLoadFolders();
+  }, []);
+
   return (
     <div className="SharedPage">
-      {folderInfo}
+      <FolderInfo
+        profileImage={profileImage}
+        ownerName={ownerName}
+        folderName={folderName}
+      />
       <div className="SharedPage-items">
-        {searchBar}
-        {cardList}
+        <SearchBar />
+        <CardList>
+          {folders?.map((link) => (
+            <ReadOnlyCard key={link?.id} {...link} />
+          ))}
+        </CardList>
       </div>
     </div>
   );
